seperating students from groups in school view
