+++ src/components/Groups/Groups.js
@@ -1,12 +1,13 @@
 import Vue from "vue";
 import Teachers from "../Teachers/Teachers.vue";
-import { createGroup, getGroups, getAllRegistrations, writeStudentToGroup, updateRegistration, removeStudentFromGroups } from "../../firebase.js";
-import { getAge } from "../../common/getAge.js";
+import Students from "./Students/Students.vue";
+import { createGroup, getGroups, getAllRegistrations } from "../../firebase.js";
 
 export default Vue.extend ({
   name: "Groups",
   components: {
-    Teachers
+    Teachers,
+    Students
   },
   props: ['users'],
   mounted() {
@@ -18,52 +19,31 @@ export default Vue.extend ({
       isLoaded: true,
       groupName: "",
       groups: [],
-      registrations: [],
-      registrationFields: [{
-        key: "name",
-        label: "Naam",
-        formatter: (value) => {
-          return `${value.first} ${value.last}`
-        }
-      }, {
-        key: "education",
-        label: "Onderwijs"
-      }, {
-        key: "group",
-        label: "Klas"
-      }, {
-        key: "birthDate",
-        label: "Leeftijd",
-        formatter: (value) => {
-          return getAge(value)
-        }
-      }],
-      selectedGroupForStudent: '',
-      selectedStudent: ''
-    }
+      registrations: []
+    };
   },
   computed: {
     state() {
-      return (this.groupName.length >= 4 && !this.groupAlreadyExists) ? true : false
+      return (this.groupName.length >= 4 && !this.groupAlreadyExists) ? true : false;
     },
     invalidFeedback() {
       if (this.groupAlreadyExists) {
-        return 'Klas bestaat al.'
+        return 'Klas bestaat al.';
       } else if (this.groupName.length > 4) {
-        return ''
+        return '';
       } else if (this.groupName.length > 0) {
-        return 'Vul minstens 4 letters in.'
+        return 'Vul minstens 4 letters in.';
       } else {
-        return 'Vul iets in.'
+        return 'Vul iets in.';
       }
     },
     validFeedback() {
-      return this.state === true ? 'Druk op klas toevoegen om de klas aan te maken!' : ''
+      return this.state === true ? 'Druk op klas toevoegen om de klas aan te maken!' : '';
     },
     groupAlreadyExists() {
       return this.groups.some((group) => {
-        return group.groupName === this.groupName
-      })
+        return group.groupName === this.groupName;
+      });
     },
   },
   methods: {
@@ -81,16 +61,6 @@ export default Vue.extend ({
     },
     async loadRegistrations() {
       this.registrations = await getAllRegistrations();
-    },
-    onRowSelectedStudent(student) {
-      this.selectedStudent = student[0]
-    },
-    async addStudentToGroup() {
-      await removeStudentFromGroups(this.selectedStudent, this.groups);
-      await writeStudentToGroup(this.selectedStudent, this.selectedGroupForStudent);
-      await updateRegistration(this.selectedStudent, this.selectedGroupForStudent);
-      await this.loadRegistrations();
-      this.selectedStudent = "";
-    },
+    }
   }
-})
+});
