+++ src/components/Groups/Groups.vue
@@ -19,26 +19,7 @@
       </b-form-group>
     </b-container>
     <Teachers v-if="groups.length && users.length" v-bind:groups="groups" v-bind:users="users" @reloadGroups="loadGroups"/>
-    <b-container v-if="registrations.length">
-      <h1>Voeg studenten toe aan klassen</h1>
-      <b-table
-        id="students"
-        ref="studentsTable"
-        striped hover selectable
-        select-mode="single"
-        :items="registrations"
-        :fields="registrationFields"
-        @row-selected="onRowSelectedStudent"
-      ></b-table>
-      <b-form v-if="selectedStudent && groups.length" inline>
-        <b-form-select v-model="selectedGroupForStudent" :options="groups" text-field="groupName">
-          <template v-slot:first>
-            <option :value="null" disabled>-- Selecteer een klas --</option>
-          </template>
-        </b-form-select>
-        <b-button @click="addStudentToGroup" size="sm">{{selectedStudent.name.first}} {{selectedStudent.name.last}} toevoegen aan klas</b-button>
-      </b-form><br>
-    </b-container>
+    <Students v-if="registrations.length" v-bind:registrations="registrations" v-bind:groups="groups"/>
   </b-container>
 </template>
